+++ packages/cater-runtime/src/config-babel.js
@@ -2,7 +2,11 @@ const clone = require('clone');
 
 const defaultOptions = {
   presets: ['env', 'react'],
-  plugins: [['add-module-exports'], ['transform-class-properties']]
+  plugins: [
+    ['add-module-exports'],
+    ['transform-class-properties'],
+    ['transform-object-rest-spread']
+  ]
 };
 
 function generate() {